+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 **Added:**
-* New GUI to adjust the position, scale and rotation of individual elements.
+* New GUI to adjust the position, scale and rotation of individual HUD elements.
   * Found next the background button in the config menu.
+* New support for Resource Packs to include a file to set default values for individual elements.
 
 **Changed:**
 * Config screen got smaller, since the positioning and scaling options are now moved into the new GUI.
 
 **Fixed:**
+* Crash when mounting any non Gummi Ship entity.
 * Client configs from the Config screen would not save.
+* SelectedXOffset config option linked to Command Menu header instead of the slot.